More compiler cleanup

+++ sharpd/sharp_zebra.c
@@ -213,7 +213,7 @@ static int route_notify_owner(ZAPI_CALLBACK_ARGS)
 			monotime(&sg.r.t_end);
 			timersub(&sg.r.t_end, &sg.r.t_start, &r);
 			zlog_debug("Installed All Items %ld.%ld", r.tv_sec,
-				   r.tv_usec);
+				   (long int)r.tv_usec);
 			handle_repeated(true);
 		}
 		break;
@@ -229,7 +229,7 @@ static int route_notify_owner(ZAPI_CALLBACK_ARGS)
 			monotime(&sg.r.t_end);
 			timersub(&sg.r.t_end, &sg.r.t_start, &r);
 			zlog_debug("Removed all Items %ld.%ld", r.tv_sec,
-				   r.tv_usec);
+				   (long int)r.tv_usec);
 			handle_repeated(false);
 		}
 		break;
